Apply suggestions from code review

+++ src/transformers/models/wav2vec2/tokenization_wav2vec2.py
@@ -97,7 +97,7 @@ WAV2VEC2_KWARGS_DOCSTRING = r"""
                 Whether or not to print more information and warnings.
 """
 
-ListOfDict = List[Dict[str, Union[float, str]]]
+ListOfDict = List[Dict[str, Union[int, str]]]
 
 
 @dataclass
